Assignment6
Implemented widget repo and adjust controller

+++ src/main/java/com/example/wbdvsu19Qinserverjava/models/Widget.java
@@ -28,7 +28,6 @@ public class Widget {
 		this.name = w.name;
 		this.id = w.id;
 		this.type = w.type;
-
 		this.ordered = w.ordered;
 		this.text = w.text;
 		this.url = w.url;
@@ -50,7 +49,7 @@ public class Widget {
 	}
 
 
-	public long getId() {
+	public Integer getId() {
 		return id;
 	}
